Move recordEvent analytics handler to composite dir
